Fix build for Debian 9
Build on Debian 9 is failing with the message:
ERROR 2002 (HY000):
Can't connect to local MySQL server through socket
'/var/run/mysqld/mysqld.sock' (2)

The reason for this message is that Debian 9
does not support aufs FS driver anymore and is using overlay2.

MySQL has issues with overlay2,
because it implements only some of POSIX standards.

The workaround is to 'touch' the files used by MySQL.

Explanation here:
https://github.com/docker/for-linux/issues/72#issuecomment-319904698
and here:
https://docs.docker.com/storage/storagedriver/overlayfs-driver/